fix: use set_password in user model
When user tried to login, there was a encription issue as the user model uses check_password to verify if the raw password matches the hashed user password stored in database. Using the model's set_password, the password is hashed using the default hash algorithm and stored in the database, fixing the issue.